// TuTurno.Frontend/src/app/models/doctorAvailability.ts
import { IDoctor } from "./doctor";

export interface IDoctorAvailability {
    id: string;
    doctorId: string;
    dayOfWeek: number; // 0=domingo, 1=lunes ... 6=sabado
    startTime: string; // Ej: "08:00"
    endTime: string; // Ej: "17:00"
    createdAt?: Date;
    updatedAt?: Date;

    doctor?: IDoctor;
}